requirements.lock.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
base58==2.1.1
blinker==1.9.0
boto3==1.38.45
botocore==1.38.45
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
ECPy==1.2.5
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.2
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.33.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.44.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
postgrest==1.1.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
storage3==0.12.0
streamlit==1.46.1
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchaudio==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0
types-Deprecated==1.2.15.20250304
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==10.4
Werkzeug==3.1.3
wrapt==1.17.2
xrpl-py==2.3.0
